Feature/task 9

+++ src/main/java/kalchenko/exception/TaskNotFoundException.java
@@ -2,7 +2,7 @@ package kalchenko.exception;
 
 public class TaskNotFoundException extends  RuntimeException{
 
-    public TaskNotFoundException(int id){
+    public TaskNotFoundException(long id){
 
         super("Could not find task " + id);
